git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi@10702 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ V2.x/trunk/SpagoBIConsoleEngine/src/it/eng/spagobi/engines/console/services/GetConsoleDataAction.java
@@ -96,7 +96,7 @@ public class GetConsoleDataAction extends AbstractConsoleEngineAction {
 			try {
 				JSONDataWriter writer = new JSONDataWriter();
 				Object resultNumber = dataStore.getMetaData().getProperty("resultNumber");
-				if(resultNumber == null) dataStore.getMetaData().setProperty("resultNumber", new Integer(1));
+				if(resultNumber == null) dataStore.getMetaData().setProperty("resultNumber", new Integer(10));
 				JSONObject dataSetJSON = (JSONObject)writer.write(dataStore);				
 				//results.put("results", dataSetJSON);
 				results = dataSetJSON;
